refactor(search-router): break search &harr; sub cycle via _search_deps.py
mypy caught the circular dep the runtime tolerates: search.py imports
sub-modules at file end, sub-modules import _get_search_daemon back
from search.py at their top. Runtime works because search.py's
namespace is populated first; mypy gives up and reports 'Cannot
determine type' on every symbol resolved through the cycle.

Extracts _get_search_daemon into _search_deps.py (zero project-side
imports beyond FastAPI). Sub-modules and search.py itself both
import from _search_deps — no cycle either at runtime or in mypy's
view. search.py re-exports _get_search_daemon so existing
test / monkeypatch call sites keep resolving via search._get_search_daemon.
